backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
mysql-connector-python>=8.0.0
pydantic>=2.10.0
python-dotenv>=1.0.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pyotp
rapidfuzz
email-validator
